refactor: changed import/export types
