+++ packages/react-ds/src/components/Button/index.stories.mdx
@@ -5,7 +5,7 @@ import { withA11y } from '@storybook/addon-a11y'
 import { knobButtonLoading, knobButtonSizes } from '@venice/core/storybook/knobs'
 import Docs from '@venice/core/docs/Button.md'
 
-import Button from './Button'
+import { Button } from '.'
 
 <Meta title="Components|Buttons" component={Button} />
