+++ packages/react-ds/src/components/Input/index.stories.mdx
@@ -5,7 +5,7 @@ import { withA11y } from '@storybook/addon-a11y'
 import { knobInputLoading, knobInputSizes } from '@venice/core/storybook/knobs'
 import Docs from '@venice/core/docs/Input.md'
 
-import Input from './Input'
+import { Input } from '.'
 
 <Meta title="Forms|Inputs" component={Input} />
