+++ packages/react-ds/src/components/Pagination/index.spec.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { render } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
-import Pagination from './Pagination'
+import { Pagination } from '.'
 
 describe('<Pagination />', () => {
   const props = {
