+++ packages/react-ds/src/components/Pagination/index.stories.mdx
@@ -6,7 +6,7 @@ import {
   Props,
 } from '@storybook/addon-docs/blocks'
 
-import Pagination from './Pagination'
+import { Pagination } from '.'
 import { withA11y } from '@storybook/addon-a11y'
 import { withKnobs } from '@storybook/addon-knobs'
 import Docs from '@venice/core/docs/Pagination.md'
