+++ packages/react-ds/src/components/Select/index.stories.mdx
@@ -2,10 +2,10 @@ import { Meta, Story, Description, Preview, Props } from '@storybook/addon-docs/
 import { withKnobs } from '@storybook/addon-knobs'
 import { withA11y } from '@storybook/addon-a11y'
 
-import Spinner from '../Spinner/Spinner'
+import { Spinner } from '../Spinner'
 import Docs from '@venice/core/docs/Select.md'
 
-import Select from './Select'
+import { Select } from '.'
 import StoreIcon from '../Icons/Store'
 
 <Meta title="Forms|Selects" component={Select} />
