+++ packages/react-ds/src/components/Spinner/index.stories.mdx
@@ -5,7 +5,7 @@ import { withA11y } from '@storybook/addon-a11y'
 import { knobSpinnerSize } from '@venice/core/storybook/knobs'
 import Docs from '@venice/core/docs/Spinner.md'
 
-import Spinner from './Spinner'
+import { Spinner } from '.'
 
 <Meta title="Components|Spinner" component={Spinner} />
