+++ packages/react-ds/src/components/Spinner/index.tsx
@@ -5,12 +5,12 @@ import classNames from 'classnames/bind'
 import { ISpinner } from '@venice/core/models'
 import styles from '@venice/styles/components/Spinner.module.scss'
 
-interface ISpinnerProps extends ISpinner {
+export interface ISpinnerProps extends ISpinner {
   /** string */
-  className: string
+  className?: string
 }
 
-const Spinner = ({
+export const Spinner = ({
   className = '',
   color = 'default',
   size = 36,
@@ -43,5 +43,3 @@ const Spinner = ({
     </svg>
   )
 }
-
-export default Spinner
