+++ packages/react-ds/src/components/SplitButton/index.stories.mdx
@@ -11,7 +11,7 @@ import { withA11y } from '@storybook/addon-a11y'
 import { knobSplitButtonOpenType, knobSplitButtonIsFitMenu, knobSplitButtonSizes } from '@venice/core/storybook/knobs'
 import Docs from '@venice/core/docs/SplitButton.md'
 
-import SplitButton from './SplitButton'
+import { SplitButton } from '.'
 
 <Meta title="Components|SplitButton" component={SplitButton} />
