+++ packages/react-ds/src/components/Stepper/index.stories.mdx
@@ -3,7 +3,7 @@ import { withA11y } from '@storybook/addon-a11y'
 
 import Docs from '@venice/core/docs/Stepper.md'
 
-import Stepper from './Stepper'
+import { Stepper } from '.'
 
 <Meta title="Components|Stepper" component={Stepper} />
