+++ packages/react-ds/src/components/Stepper/index.tsx
@@ -1,15 +1,15 @@
-import React, { FC, useCallback } from 'react'
+import React, { useCallback } from 'react'
 
 import classNames from 'classnames/bind'
 
 import { IStepper } from '@venice/core/models'
 import styles from '@venice/styles/components/Stepper.module.scss'
 
-const Stepper: FC<IStepper> = ({
+export const Stepper: React.FunctionComponent<IStepper> = ({
   steps = [],
   className,
   ...props
-}: IStepper) => {
+}) => {
   const isActive = useCallback(
     (completed: boolean): string => (completed ? styles['active'] : ''),
     []
@@ -46,5 +46,3 @@ const Stepper: FC<IStepper> = ({
     </ul>
   )
 }
-
-export default Stepper
