--- packages/react-ds/src/index.tsx
@@ -1,7 +1,0 @@
-export { default as Button } from './components/Button/Button'
-export { default as Pagination } from './components/Pagination/Pagination'
-export { default as Select } from './components/Select/Select'
-export { default as Spinner } from './components/Spinner/Spinner'
-export { default as SplitButton } from './components/SplitButton/SplitButton'
-export { default as Stepper } from './components/Stepper/Stepper'
-export { default as Input } from './components/Input/Input'
